+++ mongo/datadog_checks/mongo/config_models/instance.py
@@ -80,6 +80,8 @@ class MetricsCollectionInterval(BaseModel):
     )
     collection: Optional[int] = None
     collections_indexes_stats: Optional[int] = None
+    db_stats: Optional[int] = None
+    session_stats: Optional[int] = None
     sharded_data_distribution: Optional[int] = None
 
 
@@ -90,6 +92,7 @@ class OperationSamples(BaseModel):
     )
     collection_interval: Optional[float] = None
     enabled: Optional[bool] = None
+    explain_verbosity: Optional[str] = None
 
 
 class Schemas(BaseModel):
@@ -112,6 +115,7 @@ class SlowOperations(BaseModel):
     )
     collection_interval: Optional[float] = None
     enabled: Optional[bool] = None
+    explain_verbosity: Optional[str] = None
     max_operations: Optional[float] = None
 
 
@@ -137,6 +141,7 @@ class InstanceConfig(BaseModel):
     dbstats_tag_dbname: Optional[bool] = None
     disable_generic_tags: Optional[bool] = None
     empty_default_hostname: Optional[bool] = None
+    free_storage_metrics: Optional[bool] = None
     hosts: Optional[Union[str, tuple[str, ...]]] = None
     metric_patterns: Optional[MetricPatterns] = None
     metrics_collection_interval: Optional[MetricsCollectionInterval] = None
@@ -150,6 +155,7 @@ class InstanceConfig(BaseModel):
     server: Optional[str] = None
     service: Optional[str] = None
     slow_operations: Optional[SlowOperations] = None
+    system_database_stats: Optional[bool] = None
     tags: Optional[tuple[str, ...]] = None
     timeout: Optional[int] = None
     tls: Optional[bool] = None