+++ keda/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- towncrier release notes start -->
 
+## 1.0.1 / 2025-02-20 / Agent 7.64.0
+
+***Fixed***:
+
+* Correct the data type and collection method for 2 metrics. ([#19530](https://github.com/DataDog/integrations-core/pull/19530))
+
 ## 1.0.0 / 2024-12-26 / Agent 7.62.0
 
 ***Added***: